txn: remove cross-store write-group transactions entirely
Nothing mints one any more. `mint_txn`/`begin_txn` were the seam for committing a Postgres
transaction and a separate store's write together; a store that owns its whole retain writes
once, atomically, so it returns no handle and the SQL store never had one. Every witness/decide
call site was therefore a no-op on both sides of the branch.

Not free while unused, which is why it goes rather than stays. Twice now a path has been gated
on "did the store mint a handle" — the same question as "does the store own its rows" only while
every non-SQL store ran the protocol. Once that stopped being true the gate silently answered
False: the streaming batch write stopped calling the store's own retain, and the delta write fell
through to the Postgres branch, which takes its document lock on a `documents` row a store-owned
bank does not have, so `current_hash` came back None, the stale-chunk guard never fired, and
delta ran with no concurrency control at all. Both are fixed on this branch by re-gating on the
capability; this removes the thing they could drift against.

The maintenance loop also woke every 300s on a store-owned deployment to take a Postgres
connection, list the banks, and call a recovery sweep that returns 0 — on the deployment shape
whose point is not needing Postgres.

Gone: `MemoryTxn`; `begin_txn` / `mint_txn` / `decide_txn` / `write_txn_witness` /
`recover_pending_txns` and the `txn=` parameter on every write method; the maintenance recovery
job and its gate; `_streaming_batch_write_ext`, whose two-phase body was unreachable — the only
way in was `store_owned_for`, which delegated immediately, so the call site now calls
`_streaming_store_owned_retain` directly; and `_ExtStreamingWriteResult`, whose `aborted` was
permanently False.

Two collapses fall out. `delete_document` and `delete_memory_unit` each nested `if
store_owned_for(...)` inside the same condition with the write-group branch in an unreachable
`elif`. Consolidation's `conn.transaction()` is KEPT although its witness write is gone:
`_persist_pending_refresh_tags` is a `SELECT ... FOR UPDATE` plus an `UPDATE`, and the lock has
to span both.

One behaviour fix, because the removal makes it plain. The store-owned streaming branch set
`outbox_fired[0] = True` under a comment about "the short txn above committed the
transactional-outbox row". That path takes no connection and writes no outbox row, so the flag
suppressed the post-loop fallback and the webhook was never delivered. It no longer sets it.

`test_retain_ext_writegroup.py` becomes `test_retain_store_owned_no_connection.py`. The contract
it existed for survives the feature — a store whose rows live elsewhere must not hold the
data-plane connection across the slow write — so its two store-owned delta tests carry over
unchanged, plus a new one pinning that the streaming store-owned retain opens no connection at
all. Its witness/decide assertions, and the two in `test_consolidation_failure_isolation.py`,
describe nothing now.
