+++ hindsight-api-slim/hindsight_api/engine/memories/postgres.py
@@ -53,32 +53,28 @@ class PostgresMemories(MemoriesExtension):
         facts: list,
         document_id: str | None = None,
         defer_index: bool = False,
-        txn=None,
     ) -> list[str]:
-        # `txn` is ignored: Postgres memories live in the caller's own transaction, so the
-        # write is already atomic with it — there is no separate store to hold invisible.
         # `defer_index` is meaningless here: the INSERT that returns the ids is
         # also what indexes the facts, so there is nothing to defer.
         return await writes.insert_facts(conn=conn, ops=ops, bank_id=bank_id, facts=facts, document_id=document_id)
 
-    async def delete_facts(self, bank_id: str, unit_ids: list[str], *, txn=None) -> None:
+    async def delete_facts(self, bank_id: str, unit_ids: list[str]) -> None:
         """No-op: the caller's `memory_units` DELETE (or its FK cascade) removed them."""
 
-    async def delete_where(self, bank_id: str, predicate: DeletePredicate, txn=None) -> int:
+    async def delete_where(self, bank_id: str, predicate: DeletePredicate) -> int:
         """No-op: predicate deletes are issued as SQL by the caller that owns the transaction."""
         return 0
 
-    async def delete_document(self, *, conn, fq_table, bank_id: str, document_id: str, txn=None) -> None:
-        # `txn` ignored: Postgres memories are covered by the caller's own transaction.
+    async def delete_document(self, *, conn, fq_table, bank_id: str, document_id: str) -> None:
         await writes.delete_document(conn=conn, fq_table=fq_table, bank_id=bank_id, document_id=document_id)
 
     async def drop_bank_storage(self, bank_id: str) -> None:
         """No-op: deleting the bank cascades to its memories."""
 
-    async def delete_observations(self, *, conn, fq_table, bank_id: str, txn=None) -> None:
+    async def delete_observations(self, *, conn, fq_table, bank_id: str) -> None:
         await writes.delete_observations(conn=conn, fq_table=fq_table, bank_id=bank_id)
 
-    async def update_memories(self, bank_id: str, patches: list[MemoryPatch], txn=None) -> None:
+    async def update_memories(self, bank_id: str, patches: list[MemoryPatch]) -> None:
         """No-op: the caller's UPDATE already wrote the row it holds open."""
 
     # ------------------------------------------------------------------ recall
@@ -394,7 +390,6 @@ class PostgresMemories(MemoriesExtension):
         unit_ids: list[str],
         when: datetime | None,
         failed: bool = False,
-        txn=None,
     ) -> None:
         await reads.mark_consolidated(
             conn=conn, fq_table=fq_table, bank_id=bank_id, unit_ids=unit_ids, when=when, failed=failed
@@ -461,7 +456,7 @@ class PostgresMemories(MemoriesExtension):
 
     # ------------------------------------------------------------------ observations
 
-    async def upsert_observation(self, *, conn, bank_id: str, record, txn=None) -> None:
+    async def upsert_observation(self, *, conn, bank_id: str, record) -> None:
         """No-op: the observation was written as a `memory_units` row by the caller."""
 
     async def observations_for_sources(
@@ -523,9 +518,7 @@ class PostgresMemories(MemoriesExtension):
     async def get_archived_memory(self, *, conn, fq_table, bank_id: str, unit_id: str) -> StoredMemory | None:
         return await writes.get_archived_memory(conn=conn, fq_table=fq_table, bank_id=bank_id, unit_id=unit_id)
 
-    async def invalidate_memory(
-        self, *, conn, fq_table, bank_id: str, unit_id: str, reason: str | None, txn=None
-    ) -> bool:
+    async def invalidate_memory(self, *, conn, fq_table, bank_id: str, unit_id: str, reason: str | None) -> bool:
         return await writes.invalidate_memory(
             conn=conn, fq_table=fq_table, bank_id=bank_id, unit_id=unit_id, reason=reason
         )
@@ -535,10 +528,10 @@ class PostgresMemories(MemoriesExtension):
             conn=conn, fq_table=fq_table, bank_id=bank_id, unit_id=unit_id, reason=reason
         )
 
-    async def restore_memory(self, *, conn, fq_table, bank_id: str, unit_id: str, txn=None) -> StoredMemory | None:
+    async def restore_memory(self, *, conn, fq_table, bank_id: str, unit_id: str) -> StoredMemory | None:
         return await writes.restore_memory(conn=conn, fq_table=fq_table, bank_id=bank_id, unit_id=unit_id)
 
-    async def set_memory_embedding(self, *, conn, fq_table, bank_id: str, unit_id: str, embedding, txn=None) -> None:
+    async def set_memory_embedding(self, *, conn, fq_table, bank_id: str, unit_id: str, embedding) -> None:
         await writes.set_memory_embedding(
             conn=conn, fq_table=fq_table, bank_id=bank_id, unit_id=unit_id, embedding=embedding
         )
@@ -562,7 +555,6 @@ class PostgresMemories(MemoriesExtension):
         mentioned_at,
         entity_ids: list[str] | None,
         entity_names: list[str] | None = None,  # noqa: ARG002 — this store's registry is SQL; the host already minted+linked, so entity_ids is authoritative.
-        txn=None,
     ) -> None:
         await writes.apply_edit(
             conn=conn,
@@ -655,11 +647,8 @@ class PostgresMemories(MemoriesExtension):
         bank_id: str | None = None,
         unit_ids: list[Any],
         entity_ids: list[Any],
-        txn=None,
     ) -> None:
-        # The join is keyed by global unit id, so bank_id is not needed here. `txn` is inert: this
-        # posting is an ordinary INSERT in the caller's own transaction, which is already the unit
-        # of atomicity — there is no second store to coordinate with.
+        # The join is keyed by global unit id, so bank_id is not needed here.
         await ops.bulk_insert_unit_entities(conn, fq_table("unit_entities"), unit_ids, entity_ids)
 
     async def enqueue_relink_victims(
